requirements.txt
cmd2~=1.5.0
pip~=21.0.1
attrs~=20.3.0
wcwidth~=0.2.5
pyperclip~=1.8.2
setuptools~=54.0.0
pyreadline3~=3.3
blessed
prettytable
peewee
scapy
python-nmap
ipaddress
netifaces
psutil
tinydb
shodan
paho-mqtt
timebudget
pyshark
